Move VerifyCertChain to common library
