+++ pkg/ca/common.go
@@ -19,6 +19,7 @@ import (
 	"context"
 	"crypto"
 	"crypto/x509"
+	"errors"
 	"time"
 
 	"github.com/sigstore/fulcio/pkg/identity"
@@ -51,4 +52,60 @@ func MakeX509(ctx context.Context, principal identity.Principal, publicKey crypt
 	}
 
 	return cert, nil
+}
+
+func VerifyCertChain(certs []*x509.Certificate, signer crypto.Signer) error {
+	if len(certs) == 0 {
+		return errors.New("certificate chain must contain at least one certificate")
+	}
+
+	roots := x509.NewCertPool()
+	roots.AddCert(certs[len(certs)-1])
+
+	intermediates := x509.NewCertPool()
+	if len(certs) > 1 {
+		for _, intermediate := range certs[1 : len(certs)-1] {
+			intermediates.AddCert(intermediate)
+		}
+	}
+
+	opts := x509.VerifyOptions{
+		Roots:         roots,
+		Intermediates: intermediates,
+		KeyUsages: []x509.ExtKeyUsage{
+			x509.ExtKeyUsageCodeSigning,
+		},
+	}
+	if _, err := certs[0].Verify(opts); err != nil {
+		return err
+	}
+
+	if !certs[0].IsCA {
+		return errors.New("certificate is not a CA")
+	}
+
+	// If using an intermediate, verify that code signing extended key
+	// usage is set to satify extended key usage chainging
+	if len(certs) > 1 {
+		var hasExtKeyUsageCodeSigning bool
+		for _, extKeyUsage := range certs[0].ExtKeyUsage {
+			if extKeyUsage == x509.ExtKeyUsageCodeSigning {
+				hasExtKeyUsageCodeSigning = true
+				break
+			}
+		}
+		if !hasExtKeyUsageCodeSigning {
+			return errors.New(`certificate must have extended key usage code signing set to sign code signing certificates`)
+		}
+	}
+
+	if err := cryptoutils.EqualKeys(certs[0].PublicKey, signer.Public()); err != nil {
+		return err
+	}
+
+	if err := cryptoutils.ValidatePubKey(signer.Public()); err != nil {
+		return err
+	}
+
+	return nil
 }
