+++ pkg/ca/fileca/load.go
@@ -24,7 +24,6 @@ import (
 	"path/filepath"
 
 	"github.com/sigstore/fulcio/pkg/ca"
-	"github.com/sigstore/fulcio/pkg/ca/baseca"
 	"github.com/sigstore/sigstore/pkg/cryptoutils"
 	"go.step.sm/crypto/pemutil"
 )
@@ -58,7 +57,7 @@ func loadKeyPair(certPath, keyPath, keyPass string) (*ca.SignerCertsMutex, error
 		}
 	}
 
-	if err := baseca.VerifyCertChain(certs, key); err != nil {
+	if err := ca.VerifyCertChain(certs, key); err != nil {
 		return nil, err
 	}
